+++ horde/classes/stable/waiting_prompt.py
@@ -438,8 +438,6 @@ class ImageWaitingPrompt(WaitingPrompt):
             # These samplerS do double steps per iteration, so they're at half the speed
             # So we adjust the things to take that into account
             steps *= 2
-        if self.source_image and self.source_processing == "img2img":
-            steps *= self.gen_payload.get("denoising_strength", 0.8)
         return steps
 
     def set_job_ttl(self):